refactor: Move AI utilities to ai-utils dir
