+++ src/content-scripts/components/GenerateAIDescription/DescriptionGeneratorButton.ts
@@ -2,10 +2,10 @@ import { createHtmlElement } from "../../../utils/createHtmlElement";
 import openSaucedLogoIcon from "../../../assets/opensauced-icon.svg";
 import { getPullRequestAPIURL } from "../../../utils/urlMatchers";
 import { getDescriptionContext, isOutOfContextBounds } from "../../../utils/fetchGithubAPIData";
-import { generateDescription } from "../../../utils/aiprdescription/openai";
+import { generateDescription } from "../../../utils/ai-utils/openai";
 import { GITHUB_PR_COMMENT_TEXT_AREA_SELECTOR, SUPABASE_LOGIN_URL } from "../../../constants";
-import { insertTextAtCursor } from "../../../utils/aiprdescription/cursorPositionInsert";
-import { getAIDescriptionConfig } from "../../../utils/aiprdescription/descriptionconfig";
+import { insertTextAtCursor } from "../../../utils/ai-utils/cursorPositionInsert";
+import { getAIDescriptionConfig } from "../../../utils/ai-utils/descriptionconfig";
 import { getAuthToken, isLoggedIn } from "../../../utils/checkAuthentication";
 
 export const DescriptionGeneratorButton = () => {
