+++ src/utils/ai-utils/cursorPositionInsert.ts
