+++ src/utils/ai-utils/descriptionconfig.ts
